avoid deleting categories with related budgets data in the database

+++ src/db/category/delete.ts
@@ -3,7 +3,14 @@ import Category from './model'
 import { verifyCategoryId } from './utils'
 
 export const deleteCategory = async (userId: string, req: NextApiRequest) => {
-  await verifyCategoryId(req.query.id as string, userId)
+  const verifiedCategory = await verifyCategoryId(
+    req.query.id as string,
+    userId,
+    true
+  )
+
+  if (verifiedCategory.budgets.length > 0)
+    throw new Error('Categories with related budgets cannot be deleted')
 
   const category = await Category.findOneAndDelete({
     _id: req.query.id,
